Fix align.tex test on Travis
It has no imagers or tex binaries so the HTML only has
DimensionPlaceholder values. The real Dimension is set when images are
rendered (after the HTML has been generated); the HTML has a
placeholder string for the units, and PageTemplate.setImageData is
called by PageTemplate.processFileContent to substitute them in.

But if we have no imager, wehave no unit to substitute.

Prior to a1e0ba5, this test would have been broken when there was no
imager available. That commit flipped it.

Ideally we'd have two separate benchmark files to compare the two cases.